The fun never stops
I think all these fixes were needed due to the fact that Werror has to be disabled for stock panda to build on recent Ubuntu (newer gcc). That makes it difficult to develop/merge new code because only Travis notices these 'warningErrors' ... not a huge problem, but we should probably fix these warnings so the master branch builds on the next Ubuntu LTS version (20.04).

+++ panda/plugins/lighthouse_coverage/lighthouse_coverage.c
@@ -18,7 +18,7 @@ void before_block_exec(CPUState *cpuState, TranslationBlock *translationBlock)
 		OsiProc * process = get_current_process(cpuState);		// get a reference to the process this TranslationBlock belongs to
 	        if (process) 
 			{
-			fprintf(outputFile,"\n%s@%#018"PRIx64"", process->name, (translationBlock->pc));
+			fprintf(outputFile,"\n%s@%#018"PRIx64"", process->name, (long unsigned int)(translationBlock->pc));
 			free_osiproc(process);					// always free unused resources
 			}
 		}
